requirements.txt
### 
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.0
django-admin-interface==0.28.0
django-colorfield==0.11.0
django-mdeditor==0.1.20
django-richtextfield==1.6.1
djangorestframework==3.14.0
idna==3.6
kombu==5.3.4
Pillow==10.1.0
prompt-toolkit==3.0.43
pyTelegramBotAPI==4.14.1
python-dateutil==2.8.2
python-slugify==8.0.1
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.12
postgres==4.0
psycopg2-binary==2.9.9
psycopg2-pool==1.1
gunicorn==21.2.0